+++ config.json
@@ -1,0 +1,19 @@
+{
+  "servePort": null,
+  "externalPackages": {
+   "@angular/core": "vendor._angular_core",
+    "@angular/common": "vendor._angular_common",
+    "@angular/platform-browser": "vendor._angular_platformBrowser",
+    "@angular/platform-browser-dynamic": "vendor._angular_platformBrowserDynamic",
+    "@angular/router": "vendor._angular_router",
+    "@angular/http": "vendor._angular_http",
+    "@angular/forms": "vendor._angular_forms",
+    "rxjs": "vendor.rxjs"
+  },
+  "styles": [
+    { "src": "src/styles/app.sass", "dest": "dist/css/app.css" }
+  ],
+  "universalRoutes": [
+    "/"
+  ]
+}
