--- index.metadata.json
@@ -1,1 +1,0 @@
-[{"__symbolic":"module","version":2,"metadata":{"DatePickerModule":{"__symbolic":"class","decorators":[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"NgModule"},"arguments":[{"declarations":[{"__symbolic":"reference","module":"./src/components/ng2-datepicker.component","name":"DatePickerComponent"}],"imports":[{"__symbolic":"reference","module":"@angular/common","name":"CommonModule"},{"__symbolic":"reference","module":"@angular/forms","name":"FormsModule"},{"__symbolic":"reference","module":"ng2-slimscroll/ng2-slimscroll","name":"SlimScrollModule"}],"exports":[{"__symbolic":"reference","module":"./src/components/ng2-datepicker.component","name":"DatePickerComponent"},{"__symbolic":"reference","module":"ng2-slimscroll/ng2-slimscroll","name":"SlimScrollModule"}]}]}]}},"exports":[{"from":"./src/components/ng2-datepicker.component","export":["DatePickerOptions","DateModel"]}]},{"__symbolic":"module","exports":[{"from":"./src/components/ng2-datepicker.component","export":["DatePickerOptions","DateModel"]}],"metadata":{"DatePickerModule":{"__symbolic":"class","decorators":[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"NgModule"},"arguments":[{"declarations":[{"__symbolic":"reference","module":"./src/components/ng2-datepicker.component","name":"DatePickerComponent"}],"imports":[{"__symbolic":"reference","module":"@angular/common","name":"CommonModule"},{"__symbolic":"reference","module":"@angular/forms","name":"FormsModule"},{"__symbolic":"reference","module":"ng2-slimscroll/ng2-slimscroll","name":"SlimScrollModule"}],"exports":[{"__symbolic":"reference","module":"./src/components/ng2-datepicker.component","name":"DatePickerComponent"},{"__symbolic":"reference","module":"ng2-slimscroll/ng2-slimscroll","name":"SlimScrollModule"}]}]}]}},"version":1}]
