--- ng2-datepicker.metadata.json
@@ -1,1 +1,0 @@
-[{"__symbolic":"module","version":2,"metadata":{},"exports":[{"from":"./index"}]},{"__symbolic":"module","exports":[{"from":"./index"}],"metadata":{},"version":1}]
