+++ package.json
@@ -1,48 +1,105 @@
 {
   "name": "ng2-datepicker",
-  "version": "1.4.10",
-  "description": "Angular2 Datepicker",
-  "main": "ng2-datepicker.js",
-  "typings": "ng2-datepicker.d.ts",
+  "version": "1.5.0",
+  "description": "ng2-datepicker",
+  "scripts": {
+    "serve": "node ./scripts/index.js --serve",
+    "start": "npm run serve",
+    "roll": "node ./scripts/index.js --build",
+    "roll:prerender": "npm run roll && npm run prerender",
+    "prerender": "node ./scripts/index.js --prerender",
+    "clean": "./scripts/clean.sh",
+    "test": "karma start config/karma.conf.js",
+    "lint": "npm run tslint \"src/**/*.ts\"",
+    "tslint": "tslint",
+    "sass": "node-sass src/ng2-datepicker/ng2-datepicker.sass src/ng2-datepicker/ng2-datepicker.css",
+    "build": "npm run sass && ./node_modules/.bin/ngc -p tsconfig.build.json",
+    "prepublish": "npm run build"
+  },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/jkuri/ng2-datepicker.git"
-  },
-  "scripts": {
-    "clean": "./scripts/clean.sh",
-    "dev": "tsc --watch",
-    "lint": "tslint ./src/**/*.ts",
-    "sass": "node-sass sass/ng2-datepicker.sass src/components/ng2-datepicker.css",
-    "build": "npm run sass && ./node_modules/.bin/ngc -p tsconfig.json",
-    "prepublish": "npm run build"
+    "url": "https://github.com/jkuri/ng2-datepicker.git"
   },
   "keywords": [
     "angular",
     "angular2",
     "datepicker",
     "date",
-    "directive"
+    "component"
   ],
   "author": "Jan Kuri <jkuri88@gmail.com>",
   "license": "MIT",
-  "devDependencies": {
-    "@angular/common": "^2.3.0",
-    "@angular/compiler": "^2.3.0",
-    "@angular/compiler-cli": "^2.3.0",
-    "@angular/core": "^2.3.0",
-    "@angular/forms": "^2.3.0",
-    "@angular/platform-browser": "^2.3.0",
-    "@types/node": "^6.0.42",
-    "codelyzer": "^2.0.0-beta.1",
-    "node-sass": "^3.13.0",
-    "reflect-metadata": "^0.1.8",
-    "rxjs": "5.0.0-rc.5",
-    "tslint": "^4.0.2",
-    "typescript": "2.0.10",
-    "zone.js": "0.7.2"
+  "bugs": {
+    "url": "https://github.com/jkuri/ng2-datepicker/issues"
   },
+  "homepage": "https://github.com/jkuri/ng2-datepicker#readme",
   "dependencies": {
-    "moment": "^2.14.1",
-    "ng2-slimscroll": "^1.0.1"
+    "moment": "^2.17.1",
+    "ng2-slimscroll": "^1.2.1"
+  },
+  "devDependencies": {
+    "@angular/common": "2.4.4",
+    "@angular/compiler": "2.4.4",
+    "@angular/compiler-cli": "2.4.4",
+    "@angular/core": "2.4.4",
+    "@angular/forms": "2.4.4",
+    "@angular/http": "2.4.4",
+    "@angular/platform-browser": "2.4.4",
+    "@angular/platform-browser-dynamic": "2.4.4",
+    "@angular/platform-server": "2.4.4",
+    "@angular/router": "3.4.4",
+    "@types/browser-sync": "0.0.34",
+    "@types/chalk": "^0.4.31",
+    "@types/fs-extra": "0.0.37",
+    "@types/glob": "^5.0.30",
+    "@types/jasmine": "^2.5.41",
+    "@types/lodash": "^4.14.50",
+    "@types/node": "^7.0.4",
+    "@types/node-sass": "3.10.32",
+    "@types/optimist": "^0.0.29",
+    "@types/temp": "^0.8.29",
+    "angular2-universal": "2.1.0-rc.1",
+    "bulma": "^0.3.1",
+    "chai": "^3.5.0",
+    "chalk": "^1.1.3",
+    "chokidar": "^1.6.1",
+    "cli-progress-spinner": "0.0.2",
+    "codelyzer": "2.0.0-beta.4",
+    "core-js": "^2.4.1",
+    "fs-extra": "^2.0.0",
+    "glob": "^7.1.1",
+    "jasmine": "^2.5.3",
+    "jasmine-spec-reporter": "^3.2.0",
+    "karma": "^1.4.0",
+    "karma-chrome-launcher": "^2.0.0",
+    "karma-jasmine": "^1.1.0",
+    "karma-rollup-plugin": "0.2.4",
+    "karma-spec-reporter": "0.0.26",
+    "livereload": "^0.6.0",
+    "lodash": "^4.17.4",
+    "mocha": "^3.2.0",
+    "ng2-platform-node": "0.0.5",
+    "node-sass": "^4.3.0",
+    "open": "0.0.5",
+    "optimist": "^0.6.1",
+    "portfinder": "^1.0.12",
+    "reflect-metadata": "^0.1.9",
+    "rollup": "^0.41.4",
+    "rollup-plugin-alias": "^1.2.0",
+    "rollup-plugin-angular": "^0.4.4",
+    "rollup-plugin-buble": "^0.15.0",
+    "rollup-plugin-commonjs": "^7.0.0",
+    "rollup-plugin-node-resolve": "^2.0.0",
+    "rollup-plugin-progress": "0.2.1",
+    "rollup-plugin-serve": "^0.1.0",
+    "rollup-plugin-typescript": "^0.8.1",
+    "rollup-plugin-uglify": "^1.0.1",
+    "rxjs": "^5.0.3",
+    "rxjs-es": "^5.0.0-beta.12",
+    "temp": "^0.8.3",
+    "ts-node": "^2.0.0",
+    "tslint": "^4.3.1",
+    "typescript": "2.1.5",
+    "zone.js": "^0.7.6"
   }
 }
