+++ src/app/app.component.html
@@ -1,0 +1,11 @@
+<div class="hero is-fullheight bg-color">
+  <div class="hero-body">
+    <div class="container">
+      <div class="columns is-multiline">
+        <div class="column is-4 is-offset-4 has-text-centered">
+          <ng2-datepicker [options]="options" [(ngModel)]="date"></ng2-datepicker>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
