+++ src/app/app.component.ts
@@ -1,0 +1,15 @@
+import { Component } from '@angular/core';
+import { DatePickerOptions, DateModel } from '../../';
+
+@Component({
+  selector: 'app-root',
+  templateUrl: 'app.component.html'
+})
+export class AppComponent {
+  date: DateModel;
+  options: DatePickerOptions;
+
+  constructor() {
+    this.options = new DatePickerOptions();
+  }
+}
