+++ src/app/app.module.ts
@@ -1,0 +1,18 @@
+import { BrowserModule } from '@angular/platform-browser';
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { AppComponent } from './app.component';
+import { DatePickerModule } from '../../';
+
+@NgModule({
+  declarations: [
+    AppComponent
+  ],
+  imports: [
+    BrowserModule,
+    CommonModule,
+    DatePickerModule
+  ],
+  bootstrap: [ AppComponent ]
+})
+export class AppModule { }
