+++ src/app/index.ts
@@ -1,0 +1,3 @@
+export * from './app.component';
+export * from './app.module';
+
