+++ src/index.html
@@ -1,0 +1,16 @@
+<!doctype html>
+<html>
+<head>
+  <meta charset="utf-8">
+  <title>Angular2 Rollup Starter</title>
+  <base href="/">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <link rel="icon" type="image/x-icon" href="favicon.ico">
+  <% _.forEach(styles, function(s) { %><link rel="stylesheet" type="text/css" href="<%- s %>">
+  <% }); %>
+</head>
+<body>
+  <app-root></app-root>
+  <% _.forEach(scripts, function(s) { %><script src="<%- s %>"></script><% }); %>
+</body>
+</html>
