+++ src/main.ts
@@ -1,0 +1,4 @@
+import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { AppModule } from './app/';
+
+platformBrowserDynamic().bootstrapModule(AppModule);
