+++ src/styles/app.sass
@@ -1,0 +1,5 @@
+@import '../../node_modules/bulma/bulma'
+@import 'https://code.ionicframework.com/ionicons/2.0.1/css/ionicons.min.css'
+
+.bg-color
+  background: #F1F1E0
