+++ src/test.ts
@@ -1,0 +1,27 @@
+declare var __karma__: any;
+
+__karma__.loaded = () => { };
+
+import 'core-js/client/shim';
+import 'reflect-metadata';
+import 'zone.js/dist/zone';
+import 'zone.js/dist/long-stack-trace-zone';
+import 'zone.js/dist/async-test';
+import 'zone.js/dist/fake-async-test';
+import 'zone.js/dist/sync-test';
+import 'zone.js/dist/proxy';
+import 'zone.js/dist/jasmine-patch';
+
+import 'rxjs/Rx';
+
+import { TestBed } from '@angular/core/testing';
+import { BrowserDynamicTestingModule, platformBrowserDynamicTesting } from '@angular/platform-browser-dynamic/testing';
+
+TestBed.initTestEnvironment(
+  BrowserDynamicTestingModule,
+  platformBrowserDynamicTesting()
+);
+
+import './app/app.component.spec';
+
+__karma__.start();
