+++ src/vendor.ts
@@ -1,0 +1,24 @@
+import 'core-js/client/shim';
+import 'zone.js/dist/zone';
+import 'reflect-metadata';
+
+import * as _angular_platformBrowser from '@angular/platform-browser';
+import * as _angular_platformBrowserDynamic from '@angular/platform-browser-dynamic';
+import * as _angular_core from '@angular/core';
+import * as _angular_common from '@angular/common';
+import * as _angular_http from '@angular/http';
+import * as _angular_router from '@angular/router';
+import * as _angular_forms from '@angular/forms';
+
+import * as rxjs from 'rxjs/Rx';
+
+export default {
+  _angular_core,
+  _angular_common,
+  _angular_platformBrowser,
+  _angular_platformBrowserDynamic,
+  _angular_http,
+  _angular_router,
+  _angular_forms,
+  rxjs
+};
