+++ tsconfig.aot.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "lib": ["dom", "es6"],
+    "lib": ["dom", "es2015"],
     "target": "es5",
     "module": "es2015",
     "moduleResolution": "node",
@@ -16,10 +16,23 @@
     "noImplicitReturns": true,
     "noImplicitUseStrict": false,
     "noFallthroughCasesInSwitch": true,
-    "outDir": "./dist",
+    "outDir": "./aot",
     "rootDir": ".",
     "typeRoots": [
       "node_modules/@types"
     ]
+  },
+  "files": [
+    "src/app/app.module.ts",
+    "src/main.aot.ts"
+  ],
+  "exclude": [
+    "node_modules",
+    "aot",
+    "dist"
+  ],
+  "angularCompilerOptions": {
+    "genDir": "aot",
+    "skipMetadataEmit": true
   }
 }
